ETS 445 - Allow no result (if all files had signature error)

+++ test_fps/src/main/java/com/bosa/testfps/Main.java
@@ -660,8 +660,11 @@ public class Main implements HttpHandler {
 		MinioClient minioClient = getClient();
 
 		if (cleanupTempFiles) {
-			for(String fileToDelete : queryParams.get("toDelete").split(",")) {
-				deleteFileFromBucket(fileToDelete);
+			String filesToDelete = queryParams.get("toDelete");
+			if (!filesToDelete.equals("")) {
+				for(String fileToDelete : queryParams.get("toDelete").split(",")) {
+					deleteFileFromBucket(fileToDelete);
+				}
 			}
 
 			for(String fileToDelete : outFiles.split(",")) {
